refactoring line search computation
